Update API changes for Swan Lake
