+++ nats-native/src/main/java/org/ballerinalang/nats/connection/DefaultConnectionListener.java
@@ -37,7 +37,7 @@ public class DefaultConnectionListener implements ConnectionListener {
     private static final PrintStream console;
     private boolean printDisconnected = true;
 
-    DefaultConnectionListener() {
+    public DefaultConnectionListener() {
     }
 
     @Override
