+++ spotbugs-exclude.xml
@@ -32,7 +32,7 @@
         <Bug pattern="BC_UNCONFIRMED_CAST_OF_RETURN_VALUE" />
     </Match>
     <Match>
-        <Class name="org.ballerinalang.nats.basic.producer.Publish" />
+        <Class name="org.ballerinalang.nats.basic.client.Publish" />
         <Method name="externPublish"/>
         <Bug pattern="BC_UNCONFIRMED_CAST_OF_RETURN_VALUE" />
     </Match>
